fix: make the native addon to be context-aware
Since Electron 14, all native addons must be context-aware. Therefore,
in order to make node-tree-sitter run with Electron >= 14, it has to be
context-aware. Ref: https://github.com/electron/electron/issues/18397

In this commit, NAN_MODULE_WORKER_ENABLED is used to make the native
addon context-aware. It will allow context as the third param but will
ignore it. This change doesn't mean that the addon will be safe
to be initialised multiple times so can be run on worker threads
though. Perhaps, someone with better understanding of the library need
to implement AddEnvironmentCleanupHook to clean up resources. Ref:
https://nodejs.org/api/addons.html#context-aware-addons